@@ -15,16 +15,16 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "date-fns": "2.24.0",
+    "date-fns": "2.28.0",
     "deep-diff": "1.0.2",
-    "firebase-admin": "9.11.1",
-    "firebase-functions": "3.15.6",
-    "nodemailer": "6.6.3"
+    "firebase-admin": "10.0.1",
+    "firebase-functions": "3.16.0",
+    "nodemailer": "6.7.2"
   },
   "devDependencies": {
     "@types/deep-diff": "1.0.1",
     "@types/nodemailer": "6.4.4",
-    "firebase-functions-test": "0.3.2",
-    "typescript": "4.4.3"
+    "firebase-functions-test": "0.3.3",
+    "typescript": "4.5.4"
   }
 }
