@@ -7,67 +7,67 @@
     "functions"
   ],
   "dependencies": {
-    "@hookform/resolvers": "2.8.1",
+    "@hookform/resolvers": "2.8.5",
     "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
-    "@material-ui/lab": "4.0.0-alpha.58",
+    "@material-ui/lab": "4.0.0-alpha.60",
     "@material-ui/styles": "4.11.4",
     "@reduxjs/toolkit": "1.7.1",
-    "@sentry/react": "6.12.0",
-    "@sentry/tracing": "6.12.0",
-    "array-move": "3.0.1",
-    "autosuggest-highlight": "3.1.1",
+    "@sentry/react": "6.16.1",
+    "@sentry/tracing": "6.16.1",
+    "array-move": "4.0.0",
+    "autosuggest-highlight": "3.2.0",
     "classnames": "2.3.1",
     "clipboard-copy": "4.0.1",
-    "date-fns": "2.24.0",
+    "date-fns": "2.28.0",
     "fetch-jsonp": "1.2.1",
-    "firebase": "8.4.3",
+    "firebase": "9.6.1",
     "fontfaceobserver": "2.1.0",
     "lodash": "4.17.21",
-    "minisearch": "3.1.0",
+    "minisearch": "3.2.0",
     "normalizr": "3.6.1",
-    "prop-types": "15.7.2",
+    "prop-types": "15.8.1",
     "react": "17.0.2",
     "react-autosuggest": "10.1.0",
     "react-dom": "17.0.2",
-    "react-firebase-hooks": "3.0.4",
-    "react-hook-form": "7.15.4",
+    "react-firebase-hooks": "4.0.1",
+    "react-hook-form": "7.22.5",
     "react-redux": "7.2.6",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "6.2.1",
+    "react-scripts": "5.0.0",
     "react-sortable-hoc": "2.0.0",
-    "react-spinners": "0.10.6",
-    "react-syntax-highlighter": "15.4.4",
+    "react-spinners": "0.11.0",
+    "react-syntax-highlighter": "15.4.5",
     "redux": "4.1.2",
     "reveal.js": "4.2.1",
-    "web-vitals": "1.1.1",
+    "web-vitals": "2.1.2",
     "yup": "0.32.11"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "12.1.0",
-    "@testing-library/user-event": "13.2.1",
+    "@testing-library/jest-dom": "5.16.1",
+    "@testing-library/react": "12.1.2",
+    "@testing-library/user-event": "13.5.0",
     "@types/fontfaceobserver": "2.1.0",
-    "@types/jest": "27.0.1",
-    "@types/node": "16.9.3",
+    "@types/jest": "27.4.0",
+    "@types/node": "17.0.8",
     "@types/normalizr": "2.0.18",
-    "@types/react": "17.0.21",
-    "@types/react-dom": "17.0.9",
-    "@types/react-router-dom": "5.1.9",
+    "@types/react": "17.0.38",
+    "@types/react-dom": "17.0.11",
+    "@types/react-router-dom": "5.3.2",
     "@types/react-syntax-highlighter": "13.5.2",
-    "@types/superagent": "4.1.13",
+    "@types/superagent": "4.1.14",
     "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-prettier": "3.4.1",
-    "husky": "7.0.2",
+    "eslint-plugin-prettier": "4.0.0",
+    "husky": "7.0.4",
     "jest-fetch-mock": "3.0.3",
-    "lint-staged": "11.1.2",
-    "prettier": "2.4.1",
+    "lint-staged": "12.1.5",
+    "prettier": "2.5.1",
     "source-map-explorer": "2.5.2",
-    "typescript": "4.4.3"
+    "typescript": "4.5.4"
   },
   "scripts": {
     "postinstall": "husky install",
-    "start": "react-scripts start --no-cache",
+    "start": "react-scripts start",
     "start:server": "firebase emulators:start --import=.firebase/export",
     "build": "react-scripts build",
     "test": "react-scripts test",
@@ -156,12 +156,10 @@
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
-      "eslint --fix",
-      "git add"
+      "eslint --fix"
     ],
     "*": [
-      "prettier --write --ignore-unknown",
-      "git add"
+      "prettier --write --ignore-unknown"
     ]
   }
 }
