@@ -2,19 +2,18 @@ import React from 'react';
 
 import { useAuthState } from 'react-firebase-hooks/auth';
 
-import firebase from './firebase';
+import { auth } from './firebase';
 import AuthPage from './pages/Auth';
 import MainPage from './pages/Main';
 
 function App() {
-  const firebaseAuth = firebase.auth();
-  const [user, loading] = useAuthState(firebaseAuth);
+  const [user, loading] = useAuthState(auth);
 
   if (loading) {
     return null;
   }
 
-  return user ? <MainPage /> : <AuthPage firebaseAuth={firebaseAuth} />;
+  return user ? <MainPage /> : <AuthPage firebaseAuth={auth} />;
 }
 
 export default App;
